Use a factory to create Path instances.
This is done to allow for different implementations of the Path interface to
support reads and writes independently from a file system. Specifically the
Google App Engine does not support writes to the File System and therefore an
implementation can be provided using e.g. the data store for writes.


git-svn-id: https://svn.sosy-lab.org/software/common/branches/path-factory@350 05a39ec8-3e86-40da-8f8d-20173f296dbc