Move the configuration Builder in its own class and create a factory around it.
The need for a factory to create ConfigurationBuilder instances derives from a
problem with CPAchecker on Google App Engine. It is possible that classes create
their own Configuration instances and therefore overwrite options. However, on
App Engine a certain set of default options needs to be set at all times and
must not be overwritten. Hence, using a factory to create ConfigurationBuilder
instances makes it possible to provide a builder that will ensure this constraint.



git-svn-id: https://svn.sosy-lab.org/software/common/trunk@393 05a39ec8-3e86-40da-8f8d-20173f296dbc