Deprecated prefixes for options
Provides a "deprecated prefix" functionality for flags.
Sometimes as a result of class/package rename it is desirable
to change the prefix for the @Options annotation.
However, such a change can be dangerous as previous configuration
files will suddenly produce wrong configs (giving the warning
about unused configuration parameters in the best case).

This change allows to use a deprecated prefix, which gives a
transition period during which all configurations should be renamed.
Using a deprecated prefix will log a warning.

git-svn-id: https://svn.sosy-lab.org/software/common/trunk@528 05a39ec8-3e86-40da-8f8d-20173f296dbc