Update IntelliJ configuration files
Introduces major changes in the IDE configuration files! Let me know, if
you encounter any problems with this configuration.

We add a configuration that sets the JUnit version to 4 in the test
generator dialogue.

The major changes regard Checkstyle and Findbugs. You need to have the
IntelliJ plugins `Checkstyle-IDEA` and `Findbugs-IDEA` installed. This
configuration was tested with IntelliJ IDEA Ultimate 2016.1.1,
Checkstyle-IDEA 4.27.0, and Findbugs-IDEA 0.9.997.

The default checkstyle configuration provided in the project settings is
used as a default; for unit tests you have to select the appropriate
configuration in order to omit wrong warnings.