Change the feature of logging all used options from Configuration class.
Previously it was enabled with a config option and always wrote to
stdout. Now there is no option, but someone needs to give a PrintStream
to the Configuration class to enable the feature.

This avoids the need for injecting options into the Configuration class
itself, which makes it easier to understand, and avoids
InvalidConfigurationExceptions during creation of Configuration
instances.