Upgrade to error-prone 2.0.16.
This adds a new check which we use as a new rule:
switch statements that cover all cases of an enum,
and where all cases return or throw,
should not have a default case.
This allows better checking when the enum is extended with additional
cases.
Full explanation is at
http://errorprone.info/bugpattern/UnnecessaryDefaultInEnumSwitch

Both error-prone and Eclipse still check for the presence of a default
case if the switch does not cover all cases of the enum.