Replace FindBugs with its successor SpotBugs.
The first release of SpotBugs is backwards compatible, so almost nothing
changes, but you need to run "ant spotbugs" now.

Furthermore, we need to add a dependency on jsr305 for the
javax.annotations package, which was previously included in the
findbugs-annotations package, but this was actually bad because Guava
also has a dependency on jsr305, so we ended up with the annotations
twice on the class path.
In the future, we will probably need to migrate away from these
annotations, as both SpotBugs and Guava consider:
https://github.com/spotbugs/spotbugs/issues/130
https://github.com/spotbugs/spotbugs/pull/180
https://github.com/google/guava/issues/2960