New feature for Configuration: type converters can have different configuration options for child configurations
Some type converters themselves have options (e.g., FileTypeConverter).
Typically users create a configuration and a type converter,
and add the latter to the former.
Later, when child configurations are created (e.g., via
ConfigurationBuilder.copyFrom), the type converter is copied to the
child configuration. Now, if the type converter has options,
the type converter used by the child configuration will still use
the options from the parent configuration, which is usually undesired.

Thus we now add a feature that allows type converters to be notified
when they are copied to a child configuration,
and to replace the type-converter instance on the child configuration
with an instance that has potentially new configuration values applied.

The FileTypeConverter already makes use of this feature.
Configuration options that are not explicitly specified in the child
configuration will be copied from the parent FileTypeConverter.

Unfortunately, this requires us to add
"throws InvalidConfigurationException" to ConfiguratonBuilder.build(),
which is backwards-incompatible for source code.
However, we expect most callers of this method to handle
InvalidConfigurationException anyway because many other
Configuration-related methods throw this exception, too.