Minimum timer value is properly computed.
Minimum timer value is not always 0, but is initially Long.MAX_VALUE
so that it is computed correctly. On access,
minimum value 0 is returned if the timer still has the initial value.

Running intervals are ignored, because we do not know
yet whether they will be smaller or larger than the minimum.