Make tests pass on OpenBSD
JAVA_IO_TMPDIR.value() is "/tmp/" in OpenBSD. Adding another '/' to the end
would make the TEMP_DIR string in the FileTypeConverter class longer but
the path object will optimize the path.
Therefore "/tmp//.." is not "/tmp/.." (for a string comparison) and
"/tmp/../file" became "./file" in the FileTypeConverterTest class after a call
to CheckSafePath.