Allow BasicLogManager to be instantiated if MBeans are not available
This occurs for example if java.management is not on the module path.
Because the MBean is only an optional management interface for
dynamically changing the logging options at runtime, we can live without
it. Other errors that occur in such cases are already handled by
AbstractMBean in the same way.

Also improve the error messages of AbstractMBean while we are at it.