Improve handling of source information in configuration parser
So far, we required a string, and this string needed to be a valid path.
Otherwise, the parser would always crash.
However, the only known user of this information (except in error
messages) was FileTypeConverter, which already gracefully handled the
case of missing source information.

Furthermore, sometimes we also used dummy source strings to cover the
case that no source information is present, but this was only checked in
one case and not for example by FileTypeConverter.

So now we officially make "missing source" a supported case of our
Configuration class, remove the dummy string, and avoid the parser crash
if source is not a valid Path when loading configs from URLs.