Syntactically change default value of rootDirectory for FileTypeConverter
Due to the previous commit, rootDirectory is now used more often
(whenever no source info is present, so far it often used a dummy value).
With "." as rootDirectory we could lots of changes (e.g., in tests)
from "file" to "./file", which is equivalent.
By changing the rootDirectory from "." to the equivalent "",
we change those cases where the default value of the rootDirectory
was previously used, but these should be fewer cases.

And anyway, now it is more consistent.