Restore previous behavior of FileTypeConverter
So far, FileTypeConverter used the rootDirectory only for default values
of options, because for all non-default values the Configuration
instance actually passed a non-null source.
This differed from the documented behavior.
The second-to-last commit changed this, but for example some tests in
CPAchecker relied on the current behavior.
We can actually restore the previous behavior without difficulty,
so let's do this. The behavior is now not really consistent and
different than what was documented, but this could be improved
separately if desired.