Improve performance of exception logging in BasicLogManager
So far we used Supplier's to make the message creation lazy.
But in CPAchecker we have seen cases where the creation of the required
lambda function took some time, because this operation showed up in the
stack trace printed if CPAchecker hangs for more than 10s after a
termination request.
Now we avoid the lambda function at the cost of duplicated priority
checks.