Refactor CI config: rewrite only/except to rules
"only" and "except" are deprecated in favor of "rules".
They should still work, but there seems to be some problem with them
because sometimes the deploy-gh-pages job is triggered also on
non-default branches, where it should not run.
This can be seen in the JavaSMT repo but also if one uses the pipelines
editor on this project to simulate pipeline runs on branches.

With "rules" the bug does not seem to occur,
and it is the recommended way of writing CI configs,
so let's just use it.